+++ R/pg_list_records.R
@@ -2,7 +2,7 @@
 #'
 #' @export
 #' @param prefix A character string to specify the metadata format in OAI-PMH
-#' requests issued to the repository. The default (\code{"oai_dc"}) corresponds
+#' requests issued to the repository. The default (`oai_dc`) corresponds
 #' to the mandatory OAI unqualified Dublin Core metadata schema.
 #' @param from Character string giving datestamp to be used as lower bound for
 #' datestamp-based selective harvesting (i.e., only harvest records with
@@ -20,20 +20,21 @@
 #' asked for.
 #' @param as (character) What to return. One of "df" (for data.frame; default),
 #' "list", or "raw" (raw text)
-#' @param ... Curl debugging options passed on to \code{\link[httr]{GET}}
+#' @param ... Curl debugging options passed on to [httr::GET()]
 #' @return XML character string, data.frame, or list, depending on what
-#' requested witht the \code{as} parameter
+#' requested witht the `as` parameter
+#' @references [OAI-PMH documentation](https://www.openarchives.org/pmh/)
+#' @seealso wraps [oai::list_records()]
+#' @family oai methods
 #' @examples \dontrun{
-#' pg_list_records(set='geomound', from=Sys.Date()-10, until=Sys.Date())
+#' pg_list_records(set='citable', from=Sys.Date()-1, until=Sys.Date())
 #'
 #' # When no results found > "'noRecordsMatch'"
 #' # pg_list_records(set='geomound', from='2015-01-01', until='2015-01-01')
 #'
-#' # More examples
-#' pg_list_records(set='citable', from=Sys.Date()-1, until=Sys.Date())
-#'
 #' pg_list_records(prefix="iso19139", set='citable', from=Sys.Date()-1,
 #'   until=Sys.Date())
+#'
 #' ## FIXME - below are broken
 #' # pg_list_records(prefix="dif", set='citable', from=Sys.Date()-4,
 #' #   until=Sys.Date())